+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Changelog
 * Made filter and sort dialog scrollable, thanks @[liudongmiao](https://github.com/liudongmiao)
 * Made toggle restrictions scrollable
 * Removed setting *Restrict dangerous functions* ([issue](/../../issues/1704))
+* Added half check boxes to template ([issue](/../../issues/1706))
 * Updated simplified Chinese translation
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)